Rough and ready implementation of domain name de-compression scheme is working for QName field.
QClass, QType DataLength and Data elements are extracted from the incoming
packet.

IPv4 data is printed to the console in the rudimentary ResourceRecord class.

I'm not overly happy with the QClass/QType enum classes, there's a lot of
repeated code so they deserve a cleanup now.

TODO: Resources records of certain types e.g. NS require domain compression
logic and full packet access to do so, therefore now most of the logic is there
thinking about the design of the ResourceRecord parsing in more detail would
help.

All in all, the bare bones packet parsing work is now there in code, it just
needs a good deal of cleanup and thinking about properly!